+++ client/constants.ts
@@ -11,7 +11,7 @@ import { countersDatabase } from './content'
  * The current version of the application.
  * This should match the version in package.json
  */
-export const APP_VERSION = '0.2.1'
+export const APP_VERSION = '0.2.8t'
 
 /**
  * The maximum number of players allowed in a game.